+++ tests/testthat/test-half.life.R
@@ -158,17 +158,48 @@ test_that("half-life manual point selection", {
                         allow.tmax.in.half.life=FALSE,
                         check=FALSE)$clast.pred,
     info="manually-selected half-life respects tlast and generates a different clast.pred")
-  expect_warning(manual_blq <-
-                   pk.calc.half.life(conc=rep(0, 6),
-                                     time=c(0, 1, 2, 3, 4, 5),
-                                     manually.selected.points=TRUE,
-                                     min.hl.points=3,
-                                     tlast=20,
-                                     allow.tmax.in.half.life=FALSE,
-                                     check=FALSE),
-                 regexp="No data to manually fit for half-life (all concentrations may be 0)",
-                 fixed=TRUE,
-                 info="All BLQ with manual point selection gives a warning")
+  expect_warning(
+    manual_blq <-
+      pk.calc.half.life(
+        conc=rep(0, 6),
+        time=c(0, 1, 2, 3, 4, 5),
+        manually.selected.points=TRUE,
+        min.hl.points=3,
+        tlast=20,
+        allow.tmax.in.half.life=FALSE,
+        check=FALSE
+      ),
+    regexp="No data to manually fit for half-life (all concentrations may be 0 or excluded)",
+    fixed=TRUE,
+    info="All BLQ with manual point selection gives a warning"
+  )
   expect_true(all(is.na(unlist(manual_blq))),
               info="All BLQ with manual point selection gives all NA results")
+})
+
+test_that("two-point half-life succeeds (fix #114)", {
+  expect_equal(
+    expect_warning(
+      pk.calc.half.life(
+        conc=c(1, 0.5),
+        time=c(0, 1),
+        min.hl.points=2,
+        allow.tmax.in.half.life=TRUE,
+        check=FALSE
+      ),
+      regexp="n must be > 2 for adj.r.squared"
+    ),
+    data.frame(
+      lambda.z=log(2),
+      r.squared=1,
+      adj.r.squared=NA_real_,
+      lambda.z.time.first=0,
+      lambda.z.n.points=2,
+      clast.pred=0.5,
+      half.life=1,
+      span.ratio=1,
+      tmax=0,
+      tlast=1
+    )
+  )
 })